Improve functionality to add joins to a query in Hooks.AssocSort
- Add doc tests
- Add a join_by_assoc macro
- Add joins functionality
