+++ test/custom_hooks/assoc_sort_test.exs
@@ -1,0 +1,5 @@
+defmodule Rummage.Ecto.AssocSortTest do
+  use ExUnit.Case
+  doctest Rummage.Ecto.Hooks.AssocSort
+
+end
